+++ rspec-expectations/lib/rspec/matchers.rb
@@ -171,16 +171,6 @@ module RSpec
   #       config.include(CustomGameMatchers)
   #     end
   module Matchers
-    # @api private
-    def self.is_a_matcher?(obj)
-      return true  if ::RSpec::Matchers::BuiltIn::BaseMatcher === obj
-      return false if obj.respond_to?(:i_respond_to_everything_so_im_not_really_a_matcher)
-      return false unless obj.respond_to?(:matches?)
-
-      obj.respond_to?(:failure_message) ||
-      obj.respond_to?(:failure_message_for_should) # support legacy matchers
-    end
-
     # Passes if actual is truthy (anything but false or nil)
     def be_truthy
       BuiltIn::BeTruthy.new
@@ -275,9 +265,18 @@ module RSpec
     # You can either pass <tt>receiver</tt> and <tt>message</tt>, or a block,
     # but not both.
     #
-    # When passing a block, it must use the <tt>{ ... }</tt> format, not
-    # do/end, as <tt>{ ... }</tt> binds to the `change` method, whereas do/end
+    # When passing a block, it must use the `{ ... }` format, not
+    # do/end, as `{ ... }` binds to the `change` method, whereas do/end
     # would errantly bind to the `expect(..).to` or `expect(...).not_to` method.
+    #
+    # You can chain any of the following off of the end to specify details
+    # about the change:
+    #
+    # * `by`
+    # * `by_at_least`
+    # * `by_at_most`
+    # * `from`
+    # * `to`
     #
     # @example
     #
@@ -305,7 +304,7 @@ module RSpec
     #   string = "string"
     #   expect {
     #     string
-    #   }.not_to change { string }
+    #   }.not_to change { string }.from("string")
     #
     #   expect {
     #     person.happy_birthday
@@ -326,12 +325,13 @@ module RSpec
     #
     # == Notes
     #
-    # Evaluates <tt>receiver.message</tt> or <tt>block</tt> before and after it
-    # evaluates the block passed to <tt>expect</tt>.
+    # Evaluates `receiver.message` or `block` before and after it
+    # evaluates the block passed to `expect`.
     #
-    # <tt>expect( ... ).not_to change</tt> only supports the form with no subsequent
-    # calls to <tt>by</tt>, <tt>by_at_least</tt>, <tt>by_at_most</tt>,
-    # <tt>to</tt> or <tt>from</tt>.
+    # `expect( ... ).not_to change` supports the form that specifies `from`
+    # (which specifies what you expect the starting, unchanged value to be)
+    # but does not support forms with subsequent calls to `by`, `by_at_least`,
+    # `by_at_most` or `to`.
     def change(receiver=nil, message=nil, &block)
       BuiltIn::Change.new(receiver, message, &block)
     end
@@ -629,5 +629,15 @@ module RSpec
     end
 
     OperatorMatcher.register(Enumerable, '=~', BuiltIn::MatchArray)
+
+    # @api private
+    def self.is_a_matcher?(obj)
+      return true  if ::RSpec::Matchers::BuiltIn::BaseMatcher === obj
+      return false if obj.respond_to?(:i_respond_to_everything_so_im_not_really_a_matcher)
+      return false unless obj.respond_to?(:matches?)
+
+      obj.respond_to?(:failure_message) ||
+      obj.respond_to?(:failure_message_for_should) # support legacy matchers
+    end
   end
 end