+++ rspec-expectations/Changelog.md
@@ -13,6 +13,7 @@ Enhancements:
     * `failure_message_for_should_not` => `failure_message_when_negated`
     * `match_for_should` => `match`
     * `match_for_should_not` => `match_when_negated`
+* Improve generated descriptions from `change` matcher. (Myron Marston)
 
 Breaking Changes for 3.0.0:
 
@@ -20,10 +21,18 @@ Breaking Changes for 3.0.0:
   `Rspec` or `Spec`. (Myron Marston)
 * Remove deprecated `RSpec::Expectations.differ=`. (Myron Marston)
 * Remove support for deprecated `expect(...).should`. (Myron Marston)
+* Explicitly disallow `expect { }.not_to change { }` with `by`,
+  `by_at_least`, `by_at_most` or `to`. These have never been supported
+  but did not raise explicit errors. (Myron Marston)
+* Provide `===` rather than `==` as an alias of `matches?` for
+  all matchers.  The semantics of `===` are closer to an RSpec
+  matcher than `==`. (Myron Marston)
 
 Bug Fixes:
 
 * Fix wrong matcher descriptions with falsey expected value (yujinakayama)
+* Fix `expect { }.not_to change { }.from(x)` so that the matcher only
+  passes if the starting value is `x`. (Tyler Rick, Myron Marston)
 
 Deprecations:
 
@@ -39,6 +48,10 @@ Breaking Changes for 3.0.0:
 * Remove explicit support for 1.8.6. (Jon Rowe)
 * Remove the deprecated `be_close` matcher, preferring `be_within` instead.
   (Sam Phippen)
+* Remove the deprecated `have`, `have_at_least` and `have_at_most` matchers.
+  You can continue using those matchers through https://github.com/rspec/rspec-collection_matchers,
+  or you can rewrite your expectations with something like 
+  `expect(your_object.size).to eq(num)`. (Hugo Baraúna)
 * Rename `be_true` and `be_false` to `be_truthy` and `be_falsey`. (Sam Phippen)
 * Make `expect { }.to_not raise_error(SomeSpecificClass, message)`,
        `expect { }.to_not raise_error(SomeSpecificClass)` and